Move arbitrum cross-chain scenario into cross-chain governance scenarios file
